requirements.txt
kivy[base]
kivy_examples
zaber-motion
pyserial
